ICU-23178 Always set to bogus if allocation fails when modifying Locale.
If any memory allocation error occurs while a Locale object is being
modified this leaves the object in an unpredictable state (in contrast
to other kinds of errors, like illegal argument errors, where the same
input always leaves the object in the same state). When this happens,
the safest path forward is therefore to set the object to bogus.