ICU-20392 Simplify code by removing Locale::initBaseName().
Whatever the purpose of this helper function originally was,
subsequent refactorings have now made it obsolete.

It requires both less code and less processing to assign baseName
together with the other data fields, instead of starting over and
searching for the '@' character once again and possibly having to
move payload from Nest to Heap.